make authGenesisBuilder available for all modules
